Update for Meteor 1.3.
Note that Meteor 1.2-compatible versions of this package can still be
published using Meteor 1.2, on the 0.3.x version line.
