--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "node.js"]
-	path = node.js
-	url = https://github.com/nodejs/node.git
